Move LogicalPlan enum to expr
