+++ datafusion/core/src/logical_plan/mod.rs
@@ -23,13 +23,11 @@
 
 pub(crate) mod builder;
 mod dfschema;
-mod display;
 mod expr;
 mod expr_rewriter;
 mod expr_schema;
 mod expr_simplier;
 mod expr_visitor;
-mod extension;
 mod operators;
 pub mod plan;
 mod registry;
@@ -39,7 +37,7 @@ pub use builder::{
 };
 pub use datafusion_expr::expr_fn::binary_expr;
 pub use dfschema::{DFField, DFSchema, DFSchemaRef, ToDFSchema};
-pub use display::display_schema;
+
 pub use expr::{
     abs, acos, and, approx_distinct, approx_percentile_cont, array, ascii, asin, atan,
     avg, bit_length, btrim, call_fn, case, ceil, character_length, chr, coalesce, col,
@@ -60,14 +58,12 @@ pub use expr_rewriter::{
 pub use expr_schema::ExprSchemable;
 pub use expr_simplier::{ExprSimplifiable, SimplifyInfo};
 pub use expr_visitor::{ExprVisitable, ExpressionVisitor, Recursion};
-pub use extension::UserDefinedLogicalNode;
 pub use operators::Operator;
 pub use plan::{provider_as_source, source_as_provider};
 pub use plan::{
     CreateCatalog, CreateCatalogSchema, CreateExternalTable, CreateMemoryTable,
     CrossJoin, DropTable, EmptyRelation, FileType, JoinConstraint, JoinType, Limit,
-    LogicalPlan, Partitioning, PlanType, PlanVisitor, Repartition, TableScan, Union,
-    Values,
+    LogicalPlan, Partitioning, PlanType, PlanVisitor, Repartition, StringifiedPlan,
+    TableScan, ToStringifiedPlan, Union, UserDefinedLogicalNode, Values,
 };
-pub(crate) use plan::{StringifiedPlan, ToStringifiedPlan};
 pub use registry::FunctionRegistry;
